Step 3/IoC-URL-location/slims-scanner/requirements.txt
async-timeout==4.0.2
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==2.0.12
coloredlogs==15.0.1
decorator==5.1.1
humanfriendly==10.0
idna==3.4
importlib-metadata==4.8.3
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
packaging==21.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.6
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.0.9
PyYAML==6.0
random-user-agent==1.0.1
redis==4.3.6
requests==2.27.1
six==1.16.0
traitlets==4.3.3
typing-extensions==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
zipp==3.6.0
